% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{export_table_metadata}
\alias{export_table_metadata}
\title{Export Table Metadata}
\usage{
export_table_metadata(table_info, dbg = TRUE)
}
\arguments{
\item{table_info}{data frame containing metadata about the tables.}

\item{dbg}{logical. If \code{TRUE} (default), debug messages are shown.}
}
\description{
Export Table Metadata to CSV file
}
\keyword{internal}
