% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_regions.R
\name{get_combined_regions}
\alias{get_combined_regions}
\title{Get Regions of Combined Cells}
\usage{
get_combined_regions(content)
}
\arguments{
\item{content}{matrix of character}
}
\description{
Get regions of combined cells, i.e. cells that are not empty and that
are neighbours to each other
}
