% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_text_from_xlsx.R
\name{get_raw_text_from_xlsx}
\alias{get_raw_text_from_xlsx}
\title{Read Excel Sheets into List of Character Matrices}
\usage{
get_raw_text_from_xlsx(file, sheets = NULL, dbg = TRUE)
}
\arguments{
\item{file}{full path to Excel file}

\item{sheets}{name(s) of sheet(s) to be read, in a vector of character. If
\code{NULL}, all sheets are read.}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
This function reads all (given) sheets of one Excel file into a list of
character matrices. The idea of this function is to return the content of the
Excel sheets as pure raw text information. No type conversion is performed.
Empty rows at the beginning are not skipped which would be the default
behaviour of \code{\link[readxl]{read_xlsx}} that is called under the hood.
}
\examples{
# Path to example file
file <- system.file("extdata", "example.xlsx", package = "kwb.readxl")

# Read all sheets
sheet_text <- get_raw_text_from_xlsx(file)
}
